Server - encapsulated thread sensitive data into concurrent collections, simplified NetworkAdapter interface, removed redundant field playerId - from now on playerId == connectionId

+++ tanks/src/main/java/org/game/tanks/network/model/Command.java
@@ -4,13 +4,13 @@ public class Command extends TCPMessage {
 
   private static final long serialVersionUID = -7480487174538122148L;
 
-  private long playerToId;
+  private int playerToId;
 
-  public long getPlayerToId() {
+  public int getPlayerToId() {
     return playerToId;
   }
 
-  public void setPlayerToId(long playerToId) {
+  public void setPlayerToId(int playerToId) {
     this.playerToId = playerToId;
   }
