+++ tanks/src/main/java/org/game/tanks/network/model/command/Connect.java
@@ -5,14 +5,14 @@ import org.game.tanks.network.model.Command;
 public class Connect extends Command {
 
   private static final long serialVersionUID = -7357290118820049157L;
-  private long playerId;
+  private int playerId;
   private String playerName;
 
   public long getPlayerId() {
     return playerId;
   }
 
-  public Connect setPlayerId(long playerId) {
+  public Connect setPlayerId(int playerId) {
     this.playerId = playerId;
     return this;
   }
