+++ tanks/src/main/java/org/game/tanks/network/model/command/Disconnect.java
@@ -5,13 +5,13 @@ import org.game.tanks.network.model.Command;
 public class Disconnect extends Command {
 
   private static final long serialVersionUID = -4729948455525322010L;
-  private long playerId;
+  private int playerId;
 
-  public long getPlayerId() {
+  public int getPlayerId() {
     return playerId;
   }
 
-  public Disconnect setPlayerId(long playerId) {
+  public Disconnect setPlayerId(int playerId) {
     this.playerId = playerId;
     return this;
   }
