+++ tanks/src/main/java/org/game/tanks/network/model/event/KillEvent.java
@@ -5,13 +5,13 @@ import org.game.tanks.network.model.GameEvent;
 public class KillEvent extends GameEvent {
 
   private static final long serialVersionUID = -9018523221378618934L;
-  private long playerId;
+  private int playerId;
 
   public long getPlayerId() {
     return playerId;
   }
 
-  public KillEvent setPlayerId(long playerId) {
+  public KillEvent setPlayerId(int playerId) {
     this.playerId = playerId;
     return this;
   }
